move out archive test to separate file
