+++ tests/test_model.py
@@ -5,16 +5,6 @@ import pytest
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 import bentoml  # noqa: E402
 from bentoml.artifact import PickleArtifact  # noqa: E402
-from bentoml.archive.archiver import _validate_version_str  # noqa: E402
-
-
-def test_validate_version_str_fails():
-    with pytest.raises(ValueError):
-        _validate_version_str('44&')
-
-
-def test_validate_version_str_pass():
-    _validate_version_str('abc_123')
 
 
 class MyTestModel(object):
